--- OldFiles/DashboardHistory-OLD.svelte
@@ -1,416 +1,0 @@
-<script lang="ts">
-
-	// Imports:
-	import { getBlockExplorerLink, getTimeDisplay } from '$lib/functions';
-	import { aggregatedData, startTimestamp, endTimestamp } from '$lib/stores';
-
-	// Initializations:
-	const pageSize = 25;
-	let tabSelected: 'winners' | 'deposits' | 'delegations' = 'winners';
-	let listLength = pageSize;
-	let selectedDraw = 0;
-	let depositFilter = 0;
-
-	// Reactive Data:
-	$: draws = getDraws($startTimestamp, $endTimestamp);
-	$: winners = draws[selectedDraw].result;
-	$: deposits = getDeposits(depositFilter, $startTimestamp, $endTimestamp);
-	$: delegations = getDelegations($startTimestamp, $endTimestamp);
-
-	// Draw Info:
-	$: claimable = winners.reduce((a, b) => a + (b.claimable.reduce((a, b) => a + b, 0)), 0);
-	$: dropped = winners.reduce((a, b) => a + (b.dropped.reduce((a, b) => a + b, 0)), 0);
-	$: winning = winners.filter(wallet => wallet.claimable.length > 0).length;
-	$: when = getTimeDisplay(draws[selectedDraw].timestamp, true);
-
-	// Function to get sorted and filtered draws:
-	const getDraws = (startTime: number, endTime: number) => {
-		const allDraws = $aggregatedData.draws.data.slice();
-		const filteredDraws = allDraws.filter(draw => draw.timestamp >= startTime && draw.timestamp <= endTime).reverse();
-		return filteredDraws;
-	}
-
-	// Function to get sorted and filtered deposits:
-	const getDeposits = (filter: number, startTime: number, endTime: number) => {
-		const allDeposits = $aggregatedData.deposits.data.slice();
-		const filteredDeposits = allDeposits.filter(deposit => deposit.amount >= filter && deposit.timestamp && deposit.timestamp >= startTime && deposit.timestamp <= endTime).reverse();
-		return filteredDeposits;
-	}
-
-	// Function to get sorted and filtered delegations:
-	const getDelegations = (startTime: number, endTime: number) => {
-		const allDelegations = $aggregatedData.delegationsFunded.data.slice();
-		const filteredDelegations = allDelegations.filter(delegation => delegation.timestamp && delegation.timestamp >= startTime && delegation.timestamp <= endTime).reverse();
-		return filteredDelegations;
-	}
-	
-</script>
-
-<!-- #################################################################################################### -->
-
-<!-- History -->
-<div class="history">
-
-	<!-- Header -->
-	<div class="header">
-		<h2>PoolTogether V4 History</h2>
-		<h2 id="altHeader">History</h2>
-		{#if tabSelected === 'winners'}
-			<div id="drawSelector">
-				<span>Draw:</span>
-				<select bind:value={selectedDraw}>
-					{#each draws as draw, i}
-						<option value="{i}">{draw.draw}</option>
-					{/each}
-				</select>
-			</div>
-			<i class="icofont-list" />
-			<div id="drawInfo">
-				<div class="wrapper">
-					<h3>Draw Info</h3>
-					<span>Claimable: ${claimable.toLocaleString(undefined, { maximumFractionDigits: 0 })}</span>
-					<span>Dropped: ${dropped.toLocaleString(undefined, { maximumFractionDigits: 0 })}</span>
-					<span>Winners: {winning.toLocaleString(undefined)}</span>
-					<span>When: {when}</span>
-				</div>
-			</div>
-		{:else if tabSelected === 'deposits'}
-			<div id="depositFilter">
-				<span>Filter:</span>
-				<span class="dollar">$</span>
-				<input type="number" bind:value={depositFilter}>
-			</div>
-		{/if}
-	</div>
-
-	<!-- Tab Selection -->
-	<div class="tabs">
-		<span class:selected={tabSelected === 'winners'} on:click={() => { tabSelected = 'winners'; listLength = pageSize; }}>Winners</span>
-		<span class:selected={tabSelected === 'deposits'} on:click={() => { tabSelected = 'deposits'; listLength = pageSize; }}>Deposits</span>
-		<span class:selected={tabSelected === 'delegations'} on:click={() => { tabSelected = 'delegations'; listLength = pageSize; }}>Delegations</span>
-	</div>
-
-	<!-- Content -->
-	<div class="content">
-
-		<!-- Winners Tab -->
-		{#if tabSelected === 'winners'}
-			{#if winners.length === 0}
-				<img id="sleepingPooly" src="/images/sleeping.png" alt="Sleeping Pooly">
-				<span>No winners this draw!</span>
-			{:else}
-				{#each winners.slice(0, listLength) as winner}
-					<span class="winner listItem" class:highlightItem={winner.claimable.reduce((a, b) => a + b, 0) >= 1000}>
-						{#if winner.chain}
-							<img src="/images/{winner.chain}.svg" alt="{winner.chain}">
-						{/if}
-						<a href="{`/${winner.wallet}`}" class="wallet" title="{winner.wallet}">{winner.wallet.slice(0, 6)}…{winner.wallet.slice(-4)}</a>
-						<i class="icofont-arrow-right" />
-						<span class="prizes" title="{winner.claimable.map(value => ` $${value}`).toString().slice(1)}"><span class="label">Won </span>${winner.claimable.reduce((a, b) => a + b, 0).toLocaleString(undefined, { maximumFractionDigits: 0 })}</span>
-						{#if winner.avgBalance}
-							<span class="avgBalance">(Balance: ${winner.avgBalance.toLocaleString(undefined, { maximumFractionDigits: 0 })})</span>
-						{/if}
-					</span>
-				{/each}
-				{#if winners.length > listLength}
-					<span class="loadMore" on:click={() => listLength += pageSize}><i class="icofont-arrow-down" /> Load More <i class="icofont-arrow-down" /></span>
-				{/if}
-			{/if}
-
-		<!-- Deposits Tab -->
-		{:else if tabSelected === 'deposits'}
-			{#if deposits.length === 0}
-				<img id="sleepingPooly" src="/images/sleeping.png" alt="Sleeping Pooly">
-				<span>No deposits found...</span>
-			{:else}
-				{#each deposits.slice(0, listLength) as deposit}
-					<span class="deposit listItem" class:highlightItem={deposit.amount >= 10000}>
-						{#if deposit.chain}
-							<img src="/images/{deposit.chain}.svg" alt="{deposit.chain}">
-						{/if}
-						<a href="{`/${deposit.wallet}`}" class="wallet" title="{deposit.wallet}">{deposit.wallet.slice(0, 6)}…{deposit.wallet.slice(-4)}</a>
-						<i class="icofont-arrow-right" />
-						<span class="amount"><span class="label">Deposited </span>{deposit.amount < 0.9 ? '<$1' : `$${deposit.amount.toLocaleString(undefined, { maximumFractionDigits: 0 })}`}</span>
-						{#if deposit.timestamp}
-							<span class="time">({getTimeDisplay(deposit.timestamp)})</span>
-						{/if}
-						<a class="blockExplorerLink" href="{getBlockExplorerLink(deposit.chain, deposit.txHash)}" target="__blank"><i class="icofont-external-link" /></a>
-					</span>
-				{/each}
-				{#if deposits.length > listLength}
-					<span class="loadMore" on:click={() => listLength += pageSize}><i class="icofont-arrow-down" /> Load More <i class="icofont-arrow-down" /></span>
-				{/if}
-			{/if}
-
-		<!-- Delegations Tab -->
-		{:else if tabSelected === 'delegations'}
-			{#if delegations.length === 0}
-				<img id="sleepingPooly" src="/images/sleeping.png" alt="Sleeping Pooly">
-				<span>No delegations found...</span>
-			{:else}
-				{#each delegations.slice(0, listLength) as delegation}
-					<span class="delegation listItem" class:highlightItem={delegation.amount >= 10000}>
-						{#if delegation.chain}
-							<img src="/images/{delegation.chain}.svg" alt="{delegation.chain}">
-						{/if}
-						<a href="{`/${delegation.delegator}`}" class="wallet" title="{delegation.delegator}">{delegation.delegator.slice(0, 6)}…{delegation.delegator.slice(-4)}</a>
-						<i class="icofont-arrow-right" />
-						<span class="amount"><span class="label">Delegated </span>{delegation.amount < 0.9 ? '<$1' : `$${delegation.amount.toLocaleString(undefined, { maximumFractionDigits: 0 })}`}</span>
-						{#if delegation.timestamp}
-							<span class="time">({getTimeDisplay(delegation.timestamp)})</span>
-						{/if}
-						<a class="blockExplorerLink" href="{getBlockExplorerLink(delegation.chain, delegation.txHash)}" target="__blank"><i class="icofont-external-link" /></a>
-					</span>
-				{/each}
-				{#if delegations.length > listLength}
-					<span class="loadMore" on:click={() => listLength += pageSize}><i class="icofont-arrow-down" /> Load More <i class="icofont-arrow-down" /></span>
-				{/if}
-			{/if}
-		{/if}
-	</div>
-</div>
-
-<!-- #################################################################################################### -->
-
-<style>
-
-	div.history {
-		position: relative;
-		display: flex;
-		flex-direction: column;
-		height: calc(100vh - var(--navbar-height) - 50px - 4em);
-		width: 81em;
-		margin: 0 calc((100vw - 81em) / 2);
-		padding: 1em 0 0;
-		background: var(--dark-purple);
-		border: 2px solid var(--accent-color);
-		border-radius: 1em;
-		overflow: hidden;
-	}
-
-	div.header {
-		position: relative;
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		gap: 1em;
-		padding: 0 2em;
-	}
-
-	#altHeader {
-		display: none;
-	}
-
-	div.tabs {
-		display: flex;
-		margin-top: 1em;
-		padding: 0 1em;
-		background: var(--light-purple);
-		border-top: 2px solid var(--accent-color);
-		border-bottom: 2px solid var(--accent-color);
-		box-shadow: 0 0 20px 5px var(--dark-purple) inset;
-	}
-
-	div.tabs > span {
-		padding: 1em 2em;
-		cursor: pointer;
-		user-select: none;
-	}
-
-	div.tabs > span:not(.selected):hover {
-		background: hsla(170, 86%, 57%, 0.1);
-	}
-
-	div.tabs > span.selected {
-		color: var(--dark-purple);
-		background: var(--accent-color);
-	}
-
-	div.content {
-		display: flex;
-		flex-direction: column;
-		gap: 1em;
-		padding: 1em 2em;
-		overflow: hidden auto;
-	}
-
-	#drawSelector {
-		display: flex;
-		justify-content: end;
-		align-items: center;
-		flex: 1;
-		gap: .5em;
-	}
-
-	#drawSelector > select {
-		padding-right: .5em;
-		font-family: inherit;
-		font-size: inherit;
-		background: var(--light-purple);
-		border: none;
-		border-radius: .5em;
-	}
-
-	#drawSelector > select:focus {
-		outline: none;
-	}
-
-	#drawSelector > select > option {
-		padding-right: .5em;
-		background: var(--primary-color);
-	}
-
-	#drawInfo {
-		position: absolute;
-		inset: 1.3em 1em auto auto;
-		display: none;
-	}
-
-	#drawInfo:hover {
-		display: flex;
-	}
-
-	#drawInfo > div.wrapper {
-		display: flex;
-		flex-direction: column;
-		margin-top: 1em;
-		padding: 1em;
-		background: var(--light-purple);
-		border: 2px solid var(--accent-color);
-		border-radius: .5em;
-		box-shadow: 0 0 20px 5px var(--dark-purple) inset;
-	}
-
-	#drawInfo span:first-of-type {
-		margin-top: .5em;
-		padding-top: .5em;
-		border-top: 2px solid currentColor;
-	}
-
-	div.header i.icofont-list:hover + #drawInfo {
-		display: flex;
-	}
-
-	#depositFilter {
-		display: flex;
-		align-items: center;
-	}
-
-	#depositFilter > span.dollar {
-		margin-left: .5em;
-		padding: 0 .2em 0 .5em;
-		background: var(--light-purple);
-		border-radius: .5em 0 0 .5em;
-	}
-
-	#depositFilter > input {
-		width: 5em;
-		padding-right: .5em;
-		font-family: inherit;
-		font-size: inherit;
-		background: var(--light-purple);
-		border: none;
-		border-radius: 0 .5em .5em 0;
-	}
-
-	#depositFilter > input:focus {
-		outline: none;
-	}
-
-	#sleepingPooly {
-		width: 10em;
-		margin: 1em 0;
-	}
-
-	span.listItem {
-		display: flex;
-		align-items: center;
-		gap: 1em;
-		padding: .2em .5em;
-		background: var(--primary-color);
-		border-radius: .5em;
-	}
-
-	span.listItem:hover {
-		background-color: #4d249f90;
-	}
-
-	span.listItem > img {
-		height: 2em;
-		width: 2em;
-	}
-
-	span.listItem > a.blockExplorerLink {
-		margin-left: auto;
-		color: inherit;
-		text-decoration: none;
-	}
-
-	span.highlightItem {
-		border: 2px solid var(--secondary-color);
-	}
-
-	span.loadMore {
-		margin: 0 auto;
-		color: var(--light-purple);
-		font-size: .9em;
-		cursor: pointer;
-	}
-
-	a.wallet {
-		color: var(--secondary-color);
-		font-family: 'Courier Prime', monospace;
-		text-decoration: none;
-		cursor: pointer;
-	}
-
-	span.avgBalance {
-		color: var(--light-purple);
-		font-size: .9em;
-	}
-
-	span.time {
-		color: var(--light-purple);
-		font-size: .9em;
-	}
-
-	@media screen and (max-width: 600px) {
-		div.tabs {
-			padding: 0;
-		}
-		div.tabs > span {
-			padding: 1em;
-		}
-		span.avgBalance {
-			display: none;
-		}
-		span.time {
-			display: none;
-		}
-	}
-
-	@media screen and (max-width: 530px) {
-		span.deposit span.label {
-			display: none;
-		}
-	}
-
-	@media screen and (max-width: 460px) {
-		h2 {
-			display: none;
-		}
-		#altHeader {
-			display: block;
-		}
-		span.label {
-			display: none;
-		}
-	}
-
-	@media screen and (max-width: 410px) {
-		span.listItem {
-			gap: .5em;
-		}
-	}
-	
-</style>
