+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "pooltogether-stats",
-	"version": "2.15.0",
+	"version": "2.16.0",
 	"type": "module",
 	"private": true,
 	"scripts": {
