+++ src/lib/DashboardHistory.svelte
@@ -2,14 +2,15 @@
 
 	// Imports:
 	import { getBlockExplorerLink, getTimeDisplay, getShortWallet } from '$lib/functions';
-	import { ethData, polyData, avaxData, opData, selectedChains, loading } from '$lib/stores';
+	import { ethData, polyData, avaxData, opData, startTimestamp, endTimestamp, selectedChains, loading, advancedMode } from '$lib/stores';
 
 	// Type Imports:
 	import type { Chain, DrawData, DepositData, DelegationFundedData } from '$lib/types';
 
 	// Initializations:
 	const chains: Chain[] = ['eth', 'poly', 'avax', 'op'];
-	const pageSize = 25;
+	const pageSize: number = 25;
+	const defaultMaxTimestamp = 9_999_999_999;
 	const drawHistoryWorkerPath: string = '/workers/drawHistoryWorker.js';
 	const depositHistoryWorkerPath: string = '/workers/depositHistoryWorker.js';
 	const delegationHistoryWorkerPath: string = '/workers/delegationHistoryWorker.js';
@@ -18,19 +19,25 @@
 	let deposits: (DepositData & { chain: Chain })[] = [];
 	let delegations: (DelegationFundedData & { chain: Chain })[] = [];
 	let tabSelected: 'winners' | 'deposits' | 'delegations' = 'deposits';
-	let listLength = pageSize;
-	let selectedDraw = 0;
+	let listLength: number = pageSize;
+	let depositFilter: number = 0;
 
 	// Reactive Loading Checks:
 	$: latestDepositsLoaded = chains.every(chain => $loading[chain].basic.deposits === 'done');
 	$: latestDelegationsLoaded = chains.every(chain => $loading[chain].basic.delegations === 'done');
 	$: latestDepositsErrored = !chains.every(chain => $loading[chain].basic.deposits !== 'failed');
 	$: latestDelegationsErrored = !chains.every(chain => $loading[chain].basic.delegations !== 'failed');
+	$: depositsLoaded = chains.every(chain => $loading[chain].advanced.deposits === 'done');
+	$: delegationsLoaded = chains.every(chain => $loading[chain].advanced.delegationsFunded === 'done');
+	$: depositsErrored = !chains.every(chain => $loading[chain].advanced.deposits !== 'failed');
+	$: delegationsErrored = !chains.every(chain => $loading[chain].advanced.delegationsFunded !== 'failed');
 
 	// Reactive Data:
 	$: $selectedChains, $loading, getDraws();
-	$: $selectedChains, $loading, getDeposits();
-	$: $selectedChains, $loading, getDelegations();
+	$: $selectedChains, $loading, $startTimestamp, $endTimestamp, getDeposits();
+	$: $selectedChains, $loading, $startTimestamp, $endTimestamp, getDelegations();
+	$: firstValidDraw = draws.findIndex(draw => draw.timestamp >= $startTimestamp && draw.timestamp <= $endTimestamp);
+	$: selectedDraw = draws.length > 0 ? firstValidDraw !== -1 ? firstValidDraw : 0 : 0;
 	$: winners = draws.length > 0 ? draws[selectedDraw].result.sort((a, b) => b.claimable.reduce((a, b) => a + b, 0) - a.claimable.reduce((a, b) => a + b, 0)) : undefined;
 
 	// Draw Info:
@@ -61,13 +68,16 @@
 
 	// Function to get aggregated and sorted deposits:
 	const getDeposits = async () => {
-		if(latestDepositsLoaded) {
+		if(latestDepositsLoaded && (!$advancedMode || depositsLoaded)) {
 			await new Promise<void>((resolve, reject) => {
 				const timeout = setTimeout(() => reject(`Timed out while handling deposit history.`), workerTimeout);
 				const dataWorker = new Worker(depositHistoryWorkerPath);
 				const dataToPassToWorker = {
 					deposits: { eth: $ethData.deposits.data, poly: $polyData.deposits.data, avax: $avaxData.deposits.data, op: $opData.deposits.data },
-					selectedChains: $selectedChains
+					selectedChains: $selectedChains,
+					minTimestamp: $advancedMode ? $startTimestamp : 0,
+					maxTimestamp: $advancedMode ? $endTimestamp : defaultMaxTimestamp,
+					filter: $advancedMode ? depositFilter : 0
 				}
 				dataWorker.postMessage(dataToPassToWorker);
 				dataWorker.onmessage = (event) => {
@@ -81,13 +91,15 @@
 
 	// Function to get aggregated and sorted delegations:
 	const getDelegations = async () => {
-		if(latestDelegationsLoaded) {
+		if(latestDelegationsLoaded && (!$advancedMode || delegationsLoaded)) {
 			await new Promise<void>((resolve, reject) => {
 				const timeout = setTimeout(() => reject(`Timed out while handling delegation history.`), workerTimeout);
 				const dataWorker = new Worker(delegationHistoryWorkerPath);
 				const dataToPassToWorker = {
 					delegations: { eth: $ethData.delegationsFunded.data, poly: $polyData.delegationsFunded.data, avax: $avaxData.delegationsFunded.data, op: $opData.delegationsFunded.data },
-					selectedChains: $selectedChains
+					selectedChains: $selectedChains,
+					minTimestamp: $advancedMode ? $startTimestamp : 0,
+					maxTimestamp: $advancedMode ? $endTimestamp : defaultMaxTimestamp
 				}
 				dataWorker.postMessage(dataToPassToWorker);
 				dataWorker.onmessage = (event) => {
@@ -115,7 +127,9 @@
 				<span>Draw:</span>
 				<select bind:value={selectedDraw}>
 					{#each draws as draw, i}
-						<option value="{i}">{draw.draw}</option>
+						{#if draw.timestamp >= $startTimestamp && draw.timestamp <= $endTimestamp}
+							<option value="{i}">{draw.draw}</option>
+						{/if}
 					{/each}
 				</select>
 			</div>
@@ -128,6 +142,12 @@
 					<span>Winners: {winning.toLocaleString(undefined)}</span>
 					<span>When: {when}</span>
 				</div>
+			</div>
+		{:else if tabSelected === 'deposits' && deposits.length > 0 && $advancedMode}
+			<div id="depositFilter">
+				<span>Filter:</span>
+				<span class="dollar">$</span>
+				<input type="number" bind:value={depositFilter}>
 			</div>
 		{/if}
 	</div>
@@ -180,7 +200,7 @@
 
 		<!-- Deposits Tab -->
 		{:else if tabSelected === 'deposits'}
-			{#if latestDepositsLoaded}
+			{#if latestDepositsLoaded && (!$advancedMode || depositsLoaded)}
 				{#if deposits.length === 0}
 					<img id="sleepingPooly" src="/images/sleeping.png" alt="Sleeping Pooly">
 					<span>No deposits found...</span>
@@ -203,7 +223,7 @@
 				{/if}
 			{:else}
 				<span class="loadingInfo">
-					{#if latestDepositsErrored}
+					{#if latestDepositsErrored || ($advancedMode && depositsErrored)}
 						<img src="/images/ngmi.webp" alt="Whoops">
 						<span>Something went wrong 0.o</span>
 					{:else}
@@ -215,7 +235,7 @@
 
 		<!-- Delegations Tab -->
 		{:else if tabSelected === 'delegations'}
-			{#if latestDelegationsLoaded}
+			{#if latestDelegationsLoaded && (!advancedMode || delegationsLoaded)}
 				{#if delegations.length === 0}
 					<img id="sleepingPooly" src="/images/sleeping.png" alt="Sleeping Pooly">
 					<span>No delegations found...</span>
@@ -238,7 +258,7 @@
 				{/if}
 			{:else}
 				<span class="loadingInfo">
-					{#if latestDelegationsErrored}
+					{#if latestDelegationsErrored || ($advancedMode && delegationsErrored)}
 						<img src="/images/ngmi.webp" alt="Whoops">
 						<span>Something went wrong 0.o</span>
 					{:else}
@@ -366,6 +386,32 @@
 		margin-top: .5em;
 		padding-top: .5em;
 		border-top: 2px solid currentColor;
+	}
+
+	#depositFilter {
+		display: flex;
+		align-items: center;
+	}
+
+	#depositFilter > span.dollar {
+		margin-left: .5em;
+		padding: 0 .2em 0 .5em;
+		background: var(--light-purple);
+		border-radius: .5em 0 0 .5em;
+	}
+
+	#depositFilter > input {
+		width: 5em;
+		padding-right: .5em;
+		font-family: inherit;
+		font-size: inherit;
+		background: var(--light-purple);
+		border: none;
+		border-radius: 0 .5em .5em 0;
+	}
+
+	#depositFilter > input:focus {
+		outline: none;
 	}
 
 	div.header i.icofont-list:hover + #drawInfo {
