+++ src/lib/stores.ts
@@ -4,7 +4,7 @@ import { browser } from '$app/env';
 import { writable } from 'svelte/store';
 
 // Type Imports:
-import type { ChainStats, ChainData, SelectedChains, Loading } from '$lib/types';
+import type { ChainStats, AdvancedChainStats, ChainData, SelectedChains, Loading } from '$lib/types';
 
 /* ========================================================================================================================================================================= */
 
@@ -54,38 +54,30 @@ const defaultLoadingStatus: Loading = {
 
 /* ========================================================================================================================================================================= */
 
-// Ethereum Stats:
+// Chain Stats:
 export const ethStats = writable<ChainStats>();
-
-// Polygon Stats:
 export const polyStats = writable<ChainStats>();
-
-// Avalanche Stats:
 export const avaxStats = writable<ChainStats>();
-
-// Optimism Stats:
 export const opStats = writable<ChainStats>();
+
+// Advanced Chain Stats:
+export const ethAdvancedStats = writable<AdvancedChainStats>();
+export const polyAdvancedStats = writable<AdvancedChainStats>();
+export const avaxAdvancedStats = writable<AdvancedChainStats>();
+export const opAdvancedStats = writable<AdvancedChainStats>();
 
 /* ========================================================================================================================================================================= */
 
-// Ethereum Data:
+// Chain Data:
 export const ethData = writable<ChainData>(JSON.parse(JSON.stringify(defaultChainData)));
-
-// Polygon Data:
 export const polyData = writable<ChainData>(JSON.parse(JSON.stringify(defaultChainData)));
-
-// Avalanche Data:
 export const avaxData = writable<ChainData>(JSON.parse(JSON.stringify(defaultChainData)));
-
-// Optimism Data:
 export const opData = writable<ChainData>(JSON.parse(JSON.stringify(defaultChainData)));
 
 /* ========================================================================================================================================================================= */
 
-// Start Timestamp:
-export const startTimestamp = writable<number>(0);
-
-// End Timestamp:
+// Timestamps:
+export const startTimestamp = writable<number>(1_634_256_000);
 export const endTimestamp = writable<number>(9_999_999_999);
 
 /* ========================================================================================================================================================================= */
