+++ workers/tsconfig.json
@@ -13,6 +13,7 @@
 		"depositHistoryWorker.ts",
 		"delegationHistoryWorker.ts",
 		"drawFormattingWorker.ts",
+		"advancedStatsWorker.ts",
 		"types.ts"
 	]
 }
